# Optimize `bind_least_significant_variable`