--- a/sumcheck/src/multilinear.rs
+++ b/sumcheck/src/multilinear.rs
@@ -79,7 +79,8 @@ impl<E: FieldElement> MultiLinearPoly<E> {
                 // `(evaluations.len() / 2) - 1`. Hence, the largest value for `(i<<1)` is
                 // `evaluations.len() - 2`, and largest value for `(i<<1) + 1` is `evaluations.len() - 1`.
                 let evaluations_2i = unsafe { *self.evaluations.get_unchecked(i << 1) };
-                let evaluations_2i_plus_1 = unsafe { *self.evaluations.get_unchecked((i << 1) + 1) };
+                let evaluations_2i_plus_1 =
+                    unsafe { *self.evaluations.get_unchecked((i << 1) + 1) };
 
                 self.evaluations[i] =
                     evaluations_2i + round_challenge * (evaluations_2i_plus_1 - evaluations_2i);
@@ -91,8 +92,14 @@ impl<E: FieldElement> MultiLinearPoly<E> {
         {
             let mut result = unsafe { utils::uninit_vector(num_evals) };
             result.par_iter_mut().enumerate().for_each(|(i, ev)| {
-                *ev = self.evaluations[i << 1]
-                    + round_challenge * (self.evaluations[(i << 1) + 1] - self.evaluations[i << 1])
+                // SAFETY: This loops over [0, evaluations.len()/2). The largest value for `i` is
+                // `(evaluations.len() / 2) - 1`. Hence, the largest value for `(i<<1)` is
+                // `evaluations.len() - 2`, and largest value for `(i<<1) + 1` is `evaluations.len() - 1`.
+                let evaluations_2i = unsafe { *self.evaluations.get_unchecked(i << 1) };
+                let evaluations_2i_plus_1 =
+                    unsafe { *self.evaluations.get_unchecked((i << 1) + 1) };
+
+                *ev = evaluations_2i + round_challenge * (evaluations_2i_plus_1 - evaluations_2i);
             });
             self.evaluations = result
         }