@plafer 2024-09-17
> What would the take home message be from this optimization?

Rust inserts bound checks on every slice indexing when it's not able to prove that the indexing will not be out of bounds. For example, this will not result in any bound checks being inserted

```rs
for i in 0..v.len() {
  v[i] = 3;
}
```

since `i` clearly goes from `0` to `v.len() - 1`. While Rust can figure out on its own that some indexing expression will never be out-of-bounds (e.g. `v[i-1]` in a loop from `1` to `v.len()`), it was not able to figure out that our `self.evaluations[i << 1]` and `self.evaluations[(i << 1) + 1]` would never be out of bounds, so it inserted out-of-bounds checks.

A good tool to figure that out is [`cargo-show-asm`](https://github.com/pacak/cargo-show-asm), where you can inspect the generated assembly of a function that you specify. I'd recommend the [bounds-check-cookbook](https://github.com/Shnatsel/bounds-check-cookbook/tree/main) for a good introduction on the subject. Otherwise, you can also just run a relevant benchmark and see if it improved.

Bound checks are bad for 2 reasons: 
- it inserts a useless branch into the code
- can prevent auto-vectorization (i.e. automatic generation of SIMD instructions by the compiler)

In our case, both versions of the code vectorized the load of `self.evaluations[i << 1]` and `self.evaluations[(i<< 1) + 1]`; the only difference between the 2 generated assemblies was the removal of 2 branches (due to bound checks) in the optimized version. Since this is done in a hot loop, I suppose it ended up costing a nontrivial amount. However, I'm hoping to find cases elsewhere where inserted bound checks also prevent the use of vectorized loads/stores, in which case we should see an even bigger performance boost.

Note that since with Goldilocks we're playing with "big" `u64` values, and so the vectorization opportunities are typically limited to bundling two loads/stores together (`ldp` and `stp` on aarch64), since e.g. on aarch64 the vector register are 128 bits wide. But this is basically the basis of the argument for using 32 bit-based fields :).