@Al-Kindi-0 2024-09-18
Makes sense, thank you for the great explanation!